Refactor: MAJOR REFACTOR of post component, broke into=> Modal, PostUserAction & Metadata as new components
